Crate google-drive2

This documentation was generated from drive crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

about

get

apps

get and list

changes

get, list and watch

channels

stop

children

delete, get, insert and list

comments

delete, get, insert, list, patch and update

files

copy, delete, empty trash, get, insert, list, patch, touch, trash, untrash, update and watch

parents

delete, get, insert and list

permissions

delete, get, get id for email, insert, list, patch and update

properties

delete, get, insert, list, patch and update

realtime

get and update

replies

delete, get, insert, list, patch and update

revisions

delete, get, list, patch and update

Upload supported by ...

insert files

update realtime

update files

Download supported by ...

watch files

get files

get realtime

Subscription supported by ...

watch files

get files

watch changes

insert files

list changes

Everything else about the drive v2 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.files().watch(...).doit()
let r = hub.files().empty_trash(...).doit()
let r = hub.files().copy(...).doit()
let r = hub.files().list(...).doit()
let r = hub.files().delete(...).doit()
let r = hub.files().patch(...).doit()
let r = hub.files().update(...).doit()
let r = hub.files().insert(...).doit()
let r = hub.files().untrash(...).doit()
let r = hub.files().trash(...).doit()
let r = hub.files().touch(...).doit()
let r = hub.files().get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-drive2 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-drive2" as drive2;
use drive2::File;
use drive2::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use drive2::Drive;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Drive::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: File = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.files().patch(&req, "fileId")
             .use_content_as_indexable_text(false)
             .update_viewed_date(true)
             .timed_text_track_name("gubergren")
             .timed_text_language("sadipscing")
             .set_modified_date(true)
             .remove_parents("ea")
             .pinned(false)
             .ocr_language("justo")
             .ocr(true)
             .new_revision(true)
             .convert(true)
             .add_parents("diam")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::JsonDecodeError(err) => println!("Json failed to decode: {:?}", err),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Uploads and Downlods

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

About
An item with user information and settings.

AboutAdditionalRoleInfo
Information about supported additional roles per file type. The most specific type takes precedence.

AboutAdditionalRoleInfoRoleSets
The supported additional roles per primary role.

AboutExportFormats
The allowable export formats.

AboutFeatures
List of additional features enabled on this account.

AboutGetCallBuilder
Gets the information about the current user along with Drive API settings

AboutImportFormats
The allowable import formats.

AboutMaxUploadSizes
List of max upload sizes for each file type. The most specific type takes precedence.

AboutMethodsBuilder
A builder providing access to all methods supported on about resources. It is not used directly, but through the Drive hub.

AboutQuotaBytesByService
The amount of storage quota used by different Google services.

App
The apps resource provides a list of the apps that a user has installed, with information about each app's supported MIME types, file extensions, and other details.

AppGetCallBuilder
Gets a specific app.

AppIcons
The various icons for the app.

AppList
A list of third-party applications which the user has installed or given access to Google Drive.

AppListCallBuilder
Lists a user's installed apps.

AppMethodsBuilder
A builder providing access to all methods supported on app resources. It is not used directly, but through the Drive hub.

Change
Representation of a change to a file.

ChangeGetCallBuilder
Gets a specific change.

ChangeList
A list of changes for a user.

ChangeListCallBuilder
Lists the changes for a user.

ChangeMethodsBuilder
A builder providing access to all methods supported on change resources. It is not used directly, but through the Drive hub.

ChangeWatchCallBuilder
Subscribe to changes for a user.

Channel
An notification channel used to watch for resource changes.

ChannelMethodsBuilder
A builder providing access to all methods supported on channel resources. It is not used directly, but through the Drive hub.

ChannelStopCallBuilder
Stop watching resources through this channel

ChildList
A list of children of a file.

ChildReference
A reference to a folder's child.

ChildrenDeleteCallBuilder
Removes a child from a folder.

ChildrenGetCallBuilder
Gets a specific child reference.

ChildrenInsertCallBuilder
Inserts a file into a folder.

ChildrenListCallBuilder
Lists a folder's children.

ChildrenMethodsBuilder
A builder providing access to all methods supported on children resources. It is not used directly, but through the Drive hub.

Comment
A JSON representation of a comment on a file in Google Drive.

CommentContext
The context of the file which is being commented on.

CommentDeleteCallBuilder
Deletes a comment.

CommentGetCallBuilder
Gets a comment by ID.

CommentInsertCallBuilder
Creates a new comment on the given file.

CommentList
A JSON representation of a list of comments on a file in Google Drive.

CommentListCallBuilder
Lists a file's comments.

CommentMethodsBuilder
A builder providing access to all methods supported on comment resources. It is not used directly, but through the Drive hub.

CommentPatchCallBuilder
Updates an existing comment. This method supports patch semantics.

CommentReply
A JSON representation of a reply to a comment on a file in Google Drive.

CommentReplyList
A JSON representation of a list of replies to a comment on a file in Google Drive.

CommentUpdateCallBuilder
Updates an existing comment.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Drive
Central instance to access all Drive related resource activities

File
The metadata for a file.

FileCopyCallBuilder
Creates a copy of the specified file.

FileDeleteCallBuilder
Permanently deletes a file by ID. Skips the trash. The currently authenticated user must own the file.

FileEmptyTrashCallBuilder
Permanently deletes all of the user's trashed files.

FileGetCallBuilder
Gets a file's metadata by ID.

FileImageMediaMetadata
Metadata about image media. This will only be present for image types, and its contents will depend on what can be parsed from the image content.

FileImageMediaMetadataLocation
Geographic location information stored in the image.

FileIndexableText
Indexable text attributes for the file (can only be written)

FileInsertCallBuilder
Insert a new file.

FileLabels
A group of labels for the file.

FileList
A list of files.

FileListCallBuilder
Lists the user's files.

FileMethodsBuilder
A builder providing access to all methods supported on file resources. It is not used directly, but through the Drive hub.

FilePatchCallBuilder
Updates file metadata and/or content. This method supports patch semantics.

FileThumbnail
Thumbnail for the file. Only accepted on upload and for files that are not already thumbnailed by Google.

FileTouchCallBuilder
Set the file's updated time to the current server time.

FileTrashCallBuilder
Moves a file to the trash.

FileUntrashCallBuilder
Restores a file from the trash.

FileUpdateCallBuilder
Updates file metadata and/or content.

FileVideoMediaMetadata
Metadata about video media. This will only be present for video types.

FileWatchCallBuilder
Subscribe to changes on a file

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

ParentDeleteCallBuilder
Removes a parent from a file.

ParentGetCallBuilder
Gets a specific parent reference.

ParentInsertCallBuilder
Adds a parent folder for a file.

ParentList
A list of a file's parents.

ParentListCallBuilder
Lists a file's parents.

ParentMethodsBuilder
A builder providing access to all methods supported on parent resources. It is not used directly, but through the Drive hub.

ParentReference
A reference to a file's parent.

Permission
A permission for a file.

PermissionDeleteCallBuilder
Deletes a permission from a file.

PermissionGetCallBuilder
Gets a permission by ID.

PermissionGetIdForEmailCallBuilder
Returns the permission ID for an email address.

PermissionId
An ID for a user or group as seen in Permission items.

PermissionInsertCallBuilder
Inserts a permission for a file.

PermissionList
A list of permissions associated with a file.

PermissionListCallBuilder
Lists a file's permissions.

PermissionMethodsBuilder
A builder providing access to all methods supported on permission resources. It is not used directly, but through the Drive hub.

PermissionPatchCallBuilder
Updates a permission. This method supports patch semantics.

PermissionUpdateCallBuilder
Updates a permission.

Property
A key-value pair that is either public or private to an application.

PropertyDeleteCallBuilder
Deletes a property.

PropertyGetCallBuilder
Gets a property by its key.

PropertyInsertCallBuilder
Adds a property to a file.

PropertyList
A collection of properties, key-value pairs that are either public or private to an application.

PropertyListCallBuilder
Lists a file's properties.

PropertyMethodsBuilder
A builder providing access to all methods supported on property resources. It is not used directly, but through the Drive hub.

PropertyPatchCallBuilder
Updates a property. This method supports patch semantics.

PropertyUpdateCallBuilder
Updates a property.

RealtimeGetCallBuilder
Exports the contents of the Realtime API data model associated with this file as JSON.

RealtimeMethodsBuilder
A builder providing access to all methods supported on realtime resources. It is not used directly, but through the Drive hub.

RealtimeUpdateCallBuilder
Overwrites the Realtime API data model associated with this file with the provided JSON data model.

ReplyDeleteCallBuilder
Deletes a reply.

ReplyGetCallBuilder
Gets a reply.

ReplyInsertCallBuilder
Creates a new reply to the given comment.

ReplyListCallBuilder
Lists all of the replies to a comment.

ReplyMethodsBuilder
A builder providing access to all methods supported on reply resources. It is not used directly, but through the Drive hub.

ReplyPatchCallBuilder
Updates an existing reply. This method supports patch semantics.

ReplyUpdateCallBuilder
Updates an existing reply.

Revision
A revision of a file.

RevisionDeleteCallBuilder
Removes a revision.

RevisionGetCallBuilder
Gets a specific revision.

RevisionList
A list of revisions of a file.

RevisionListCallBuilder
Lists a file's revisions.

RevisionMethodsBuilder
A builder providing access to all methods supported on revision resources. It is not used directly, but through the Drive hub.

RevisionPatchCallBuilder
Updates a revision. This method supports patch semantics.

RevisionUpdateCallBuilder
Updates a revision.

User
The JSON template for a user.

UserPicture
The user's profile picture.

Enums

Result
A universal result type used as return for all action method results.

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

ResponseResult
Identifies types which are used in API responses.

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.