Module google_drive2::api

Structs

About
An item with user information and settings.

AboutAdditionalRoleInfo
Information about supported additional roles per file type. The most specific type takes precedence.

AboutAdditionalRoleInfoRoleSets
The supported additional roles per primary role.

AboutDriveThemes
A list of themes that are supported for shared drives.

AboutExportFormats
The allowable export formats.

AboutFeatures
List of additional features enabled on this account.

AboutGetCall
Gets the information about the current user along with Drive API settings

AboutImportFormats
The allowable import formats.

AboutMaxUploadSizes
List of max upload sizes for each file type. The most specific type takes precedence.

AboutMethods
A builder providing access to all methods supported on about resources. It is not used directly, but through the DriveHub hub.

AboutQuotaBytesByService
The amount of storage quota used by different Google services.

AboutTeamDriveThemes
Deprecated - use driveThemes instead.

App
The apps resource provides a list of the apps that a user has installed, with information about each app's supported MIME types, file extensions, and other details.

AppGetCall
Gets a specific app.

AppIcons
The various icons for the app.

AppList
A list of third-party applications which the user has installed or given access to Google Drive.

AppListCall
Lists a user's installed apps.

AppMethods
A builder providing access to all methods supported on app resources. It is not used directly, but through the DriveHub hub.

Change
Representation of a change to a file or shared drive.

ChangeGetCall
Deprecated - Use changes.getStartPageToken and changes.list to retrieve recent changes.

ChangeGetStartPageTokenCall
Gets the starting pageToken for listing future changes.

ChangeList
A list of changes for a user.

ChangeListCall
Lists the changes for a user or shared drive.

ChangeMethods
A builder providing access to all methods supported on change resources. It is not used directly, but through the DriveHub hub.

ChangeWatchCall
Subscribe to changes for a user.

Channel
An notification channel used to watch for resource changes.

ChannelMethods
A builder providing access to all methods supported on channel resources. It is not used directly, but through the DriveHub hub.

ChannelStopCall
Stop watching resources through this channel

ChildList
A list of children of a file.

ChildReference
A reference to a folder's child.

ChildrenDeleteCall
Removes a child from a folder.

ChildrenGetCall
Gets a specific child reference.

ChildrenInsertCall
Inserts a file into a folder.

ChildrenListCall
Lists a folder's children.

ChildrenMethods
A builder providing access to all methods supported on children resources. It is not used directly, but through the DriveHub hub.

Comment
A comment on a file in Google Drive.

CommentContext
The context of the file which is being commented on.

CommentDeleteCall
Deletes a comment.

CommentGetCall
Gets a comment by ID.

CommentInsertCall
Creates a new comment on the given file.

CommentList
A list of comments on a file in Google Drive.

CommentListCall
Lists a file's comments.

CommentMethods
A builder providing access to all methods supported on comment resources. It is not used directly, but through the DriveHub hub.

CommentPatchCall
Updates an existing comment.

CommentReply
A comment on a file in Google Drive.

CommentReplyList
A list of replies to a comment on a file in Google Drive.

CommentUpdateCall
Updates an existing comment.

ContentRestriction
A restriction for accessing the content of the file.

Drive
Representation of a shared drive.

DriveBackgroundImageFile
An image file and cropping parameters from which a background image for this shared drive is set. This is a write only field; it can only be set on drive.drives.update requests that don't set themeId. When specified, all fields of the backgroundImageFile must be set.

DriveCapabilities
Capabilities the current user has on this shared drive.

DriveDeleteCall
Permanently deletes a shared drive for which the user is an organizer. The shared drive cannot contain any untrashed items.

DriveGetCall
Gets a shared drive's metadata by ID.

DriveHideCall
Hides a shared drive from the default view.

DriveHub
Central instance to access all DriveHub related resource activities

DriveInsertCall
Creates a new shared drive.

DriveList
A list of shared drives.

DriveListCall
Lists the user's shared drives.

DriveMethods
A builder providing access to all methods supported on drive resources. It is not used directly, but through the DriveHub hub.

DriveRestrictions
A set of restrictions that apply to this shared drive or items inside this shared drive.

DriveUnhideCall
Restores a shared drive to the default view.

DriveUpdateCall
Updates the metadata for a shared drive.

File
The metadata for a file.

FileCapabilities
Capabilities the current user has on this file. Each capability corresponds to a fine-grained action that a user may take.

FileCopyCall
Creates a copy of the specified file. Folders cannot be copied.

FileDeleteCall
Permanently deletes a file by ID. Skips the trash. The currently authenticated user must own the file or be an organizer on the parent for shared drive files.

FileEmptyTrashCall
Permanently deletes all of the user's trashed files.

FileExportCall
Exports a Google Doc to the requested MIME type and returns the exported content. Please note that the exported content is limited to 10MB.

FileGenerateIdCall
Generates a set of file IDs which can be provided in insert or copy requests.

FileGetCall
Gets a file's metadata by ID.

FileImageMediaMetadata
Metadata about image media. This will only be present for image types, and its contents will depend on what can be parsed from the image content.

FileImageMediaMetadataLocation
Geographic location information stored in the image.

FileIndexableText
Indexable text attributes for the file (can only be written)

FileInsertCall
Insert a new file.

FileLabels
A group of labels for the file.

FileList
A list of files.

FileListCall
Lists the user's files.

FileMethods
A builder providing access to all methods supported on file resources. It is not used directly, but through the DriveHub hub.

FilePatchCall
Updates file metadata and/or content. This method supports patch semantics.

FileShortcutDetails
Shortcut file details. Only populated for shortcut files, which have the mimeType field set to application/vnd.google-apps.shortcut.

FileThumbnail
A thumbnail for the file. This will only be used if a standard thumbnail cannot be generated.

FileTouchCall
Set the file's updated time to the current server time.

FileTrashCall
Moves a file to the trash. The currently authenticated user must own the file or be at least a fileOrganizer on the parent for shared drive files. Only the owner may trash a file. The trashed item is excluded from all files.list responses returned for any user who does not own the file. However, all users with access to the file can see the trashed item metadata in an API response. All users with access can copy, download, export, and share the file.

FileUntrashCall
Restores a file from the trash. The currently authenticated user must own the file or be at least a fileOrganizer on the parent for shared drive files. Only the owner may untrash a file.

FileUpdateCall
Updates file metadata and/or content.

FileVideoMediaMetadata
Metadata about video media. This will only be present for video types.

FileWatchCall
Subscribe to changes on a file

GeneratedIds
A list of generated IDs which can be provided in insert requests

ParentDeleteCall
Removes a parent from a file.

ParentGetCall
Gets a specific parent reference.

ParentInsertCall
Adds a parent folder for a file.

ParentList
A list of a file's parents.

ParentListCall
Lists a file's parents.

ParentMethods
A builder providing access to all methods supported on parent resources. It is not used directly, but through the DriveHub hub.

ParentReference
A reference to a file's parent.

Permission
A permission for a file.

PermissionDeleteCall
Deletes a permission from a file or shared drive.

PermissionGetCall
Gets a permission by ID.

PermissionGetIdForEmailCall
Returns the permission ID for an email address.

PermissionId
An ID for a user or group as seen in Permission items.

PermissionInsertCall
Inserts a permission for a file or shared drive.

PermissionList
A list of permissions associated with a file.

PermissionListCall
Lists a file's or shared drive's permissions.

PermissionMethods
A builder providing access to all methods supported on permission resources. It is not used directly, but through the DriveHub hub.

PermissionPatchCall
Updates a permission using patch semantics.

PermissionPermissionDetails
Details of whether the permissions on this shared drive item are inherited or directly on this item. This is an output-only field which is present only for shared drive items.

PermissionTeamDrivePermissionDetails
Deprecated - use permissionDetails instead.

PermissionUpdateCall
Updates a permission.

Property
A key-value pair attached to a file that is either public or private to an application. The following limits apply to file properties:

PropertyDeleteCall
Deletes a property.

PropertyGetCall
Gets a property by its key.

PropertyInsertCall
Adds a property to a file, or updates it if it already exists.

PropertyList
A collection of properties, key-value pairs that are either public or private to an application.

PropertyListCall
Lists a file's properties.

PropertyMethods
A builder providing access to all methods supported on property resources. It is not used directly, but through the DriveHub hub.

PropertyPatchCall
Updates a property.

PropertyUpdateCall
Updates a property.

ReplyDeleteCall
Deletes a reply.

ReplyGetCall
Gets a reply.

ReplyInsertCall
Creates a new reply to the given comment.

ReplyListCall
Lists all of the replies to a comment.

ReplyMethods
A builder providing access to all methods supported on reply resources. It is not used directly, but through the DriveHub hub.

ReplyPatchCall
Updates an existing reply.

ReplyUpdateCall
Updates an existing reply.

Revision
A revision of a file.

RevisionDeleteCall
Permanently deletes a file version. You can only delete revisions for files with binary content, like images or videos. Revisions for other files, like Google Docs or Sheets, and the last remaining file version can't be deleted.

RevisionGetCall
Gets a specific revision.

RevisionList
A list of revisions of a file.

RevisionListCall
Lists a file's revisions.

RevisionMethods
A builder providing access to all methods supported on revision resources. It is not used directly, but through the DriveHub hub.

RevisionPatchCall
Updates a revision.

RevisionUpdateCall
Updates a revision.

StartPageToken
There is no detailed description.

TeamDrive
Deprecated: use the drive collection instead.

TeamDriveBackgroundImageFile
An image file and cropping parameters from which a background image for this Team Drive is set. This is a write only field; it can only be set on drive.teamdrives.update requests that don't set themeId. When specified, all fields of the backgroundImageFile must be set.

TeamDriveCapabilities
Capabilities the current user has on this Team Drive.

TeamDriveList
A list of Team Drives.

TeamDriveRestrictions
A set of restrictions that apply to this Team Drive or items inside this Team Drive.

TeamdriveDeleteCall
Deprecated use drives.delete instead.

TeamdriveGetCall
Deprecated use drives.get instead.

TeamdriveInsertCall
Deprecated use drives.insert instead.

TeamdriveListCall
Deprecated use drives.list instead.

TeamdriveMethods
A builder providing access to all methods supported on teamdrive resources. It is not used directly, but through the DriveHub hub.

TeamdriveUpdateCall
Deprecated use drives.update instead.

User
Information about a Drive user.

UserPicture
The user's profile picture.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.