Struct google_drive2::api::About

pub struct About {
    pub additional_role_info: Option<Vec<AboutAdditionalRoleInfo>>,
    pub can_create_drives: Option<bool>,
    pub can_create_team_drives: Option<bool>,
    pub domain_sharing_policy: Option<String>,
    pub drive_themes: Option<Vec<AboutDriveThemes>>,
    pub etag: Option<String>,
    pub export_formats: Option<Vec<AboutExportFormats>>,
    pub features: Option<Vec<AboutFeatures>>,
    pub folder_color_palette: Option<Vec<String>>,
    pub import_formats: Option<Vec<AboutImportFormats>>,
    pub is_current_app_installed: Option<bool>,
    pub kind: Option<String>,
    pub language_code: Option<String>,
    pub largest_change_id: Option<String>,
    pub max_upload_sizes: Option<Vec<AboutMaxUploadSizes>>,
    pub name: Option<String>,
    pub permission_id: Option<String>,
    pub quota_bytes_by_service: Option<Vec<AboutQuotaBytesByService>>,
    pub quota_bytes_total: Option<String>,
    pub quota_bytes_used: Option<String>,
    pub quota_bytes_used_aggregate: Option<String>,
    pub quota_bytes_used_in_trash: Option<String>,
    pub quota_type: Option<String>,
    pub remaining_change_ids: Option<String>,
    pub root_folder_id: Option<String>,
    pub self_link: Option<String>,
    pub team_drive_themes: Option<Vec<AboutTeamDriveThemes>>,
    pub user: Option<User>,
}

An item with user information and settings.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get about (response)

Fields
additional_role_info: Option<Vec<AboutAdditionalRoleInfo>>

Information about supported additional roles per file type. The most specific type takes precedence.

can_create_drives: Option<bool>

Whether the user can create shared drives.

can_create_team_drives: Option<bool>

Deprecated - use canCreateDrives instead.

domain_sharing_policy: Option<String>

The domain sharing policy for the current user. Possible values are:

allowed

allowedWithWarning

incomingOnly

disallowed

drive_themes: Option<Vec<AboutDriveThemes>>

A list of themes that are supported for shared drives.

etag: Option<String>

The ETag of the item.

export_formats: Option<Vec<AboutExportFormats>>

The allowable export formats.

features: Option<Vec<AboutFeatures>>

List of additional features enabled on this account.

folder_color_palette: Option<Vec<String>>

The palette of allowable folder colors as RGB hex strings.

import_formats: Option<Vec<AboutImportFormats>>

The allowable import formats.

is_current_app_installed: Option<bool>

A boolean indicating whether the authenticated app is installed by the authenticated user.

kind: Option<String>

This is always drive#about.

language_code: Option<String>

The user's language or locale code, as defined by BCP 47, with some extensions from Unicode's LDML format (http://www.unicode.org/reports/tr35/).

largest_change_id: Option<String>

The largest change id.

max_upload_sizes: Option<Vec<AboutMaxUploadSizes>>

List of max upload sizes for each file type. The most specific type takes precedence.

name: Option<String>

The name of the current user.

permission_id: Option<String>

The current user's ID as visible in the permissions collection.

quota_bytes_by_service: Option<Vec<AboutQuotaBytesByService>>

The amount of storage quota used by different Google services.

quota_bytes_total: Option<String>

The total number of quota bytes. This is only relevant when quotaType is LIMITED.

quota_bytes_used: Option<String>

The number of quota bytes used by Google Drive.

quota_bytes_used_aggregate: Option<String>

The number of quota bytes used by all Google apps (Drive, Picasa, etc.).

quota_bytes_used_in_trash: Option<String>

The number of quota bytes used by trashed items.

quota_type: Option<String>

The type of the user's storage quota. Possible values are:

LIMITED

UNLIMITED

remaining_change_ids: Option<String>

The number of remaining change ids, limited to no more than 2500.

root_folder_id: Option<String>

The id of the root folder.

self_link: Option<String>

A link back to this item.

team_drive_themes: Option<Vec<AboutTeamDriveThemes>>

Deprecated - use driveThemes instead.

user: Option<User>

The authenticated user.