Struct google_drive2::api::FileMethods

pub struct FileMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on file resources. It is not used directly, but through the DriveHub hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_drive2 as drive2;
 
use std::default::Default;
use oauth2;
use drive2::DriveHub;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = DriveHub::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `copy(...)`, `delete(...)`, `empty_trash(...)`, `export(...)`, `generate_ids(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)`, `touch(...)`, `trash(...)`, `untrash(...)`, `update(...)` and `watch(...)`
// to build up your call.
let rb = hub.files();

Implementations

impl<'a, C> FileMethods<'a, C>

pub fn copy(&self, request: File, file_id: &str) -> FileCopyCall<'a, C>

Create a builder to help you perform the following task:

Creates a copy of the specified file. Folders cannot be copied.

Arguments

request - No description provided.

fileId - The ID of the file to copy.

pub fn delete(&self, file_id: &str) -> FileDeleteCall<'a, C>

Create a builder to help you perform the following task:

Permanently deletes a file by ID. Skips the trash. The currently authenticated user must own the file or be an organizer on the parent for shared drive files.

Arguments

fileId - The ID of the file to delete.

pub fn empty_trash(&self) -> FileEmptyTrashCall<'a, C>

Create a builder to help you perform the following task:

Permanently deletes all of the user's trashed files.

pub fn export(&self, file_id: &str, mime_type: &str) -> FileExportCall<'a, C>

Create a builder to help you perform the following task:

Exports a Google Doc to the requested MIME type and returns the exported content. Please note that the exported content is limited to 10MB.

Arguments

fileId - The ID of the file.

mimeType - The MIME type of the format requested for this export.

pub fn generate_ids(&self) -> FileGenerateIdCall<'a, C>

Create a builder to help you perform the following task:

Generates a set of file IDs which can be provided in insert or copy requests.

pub fn get(&self, file_id: &str) -> FileGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a file's metadata by ID.

Arguments

fileId - The ID for the file in question.

pub fn insert(&self, request: File) -> FileInsertCall<'a, C>

Create a builder to help you perform the following task:

Insert a new file.

Arguments

request - No description provided.

pub fn list(&self) -> FileListCall<'a, C>

Create a builder to help you perform the following task:

Lists the user's files.

pub fn patch(&self, request: File, file_id: &str) -> FilePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates file metadata and/or content. This method supports patch semantics.

Arguments

request - No description provided.

fileId - The ID of the file to update.

pub fn touch(&self, file_id: &str) -> FileTouchCall<'a, C>

Create a builder to help you perform the following task:

Set the file's updated time to the current server time.

Arguments

fileId - The ID of the file to update.

pub fn trash(&self, file_id: &str) -> FileTrashCall<'a, C>

Create a builder to help you perform the following task:

Moves a file to the trash. The currently authenticated user must own the file or be at least a fileOrganizer on the parent for shared drive files. Only the owner may trash a file. The trashed item is excluded from all files.list responses returned for any user who does not own the file. However, all users with access to the file can see the trashed item metadata in an API response. All users with access can copy, download, export, and share the file.

Arguments

fileId - The ID of the file to trash.

pub fn untrash(&self, file_id: &str) -> FileUntrashCall<'a, C>

Create a builder to help you perform the following task:

Restores a file from the trash. The currently authenticated user must own the file or be at least a fileOrganizer on the parent for shared drive files. Only the owner may untrash a file.

Arguments

fileId - The ID of the file to untrash.

pub fn update(&self, request: File, file_id: &str) -> FileUpdateCall<'a, C>

Create a builder to help you perform the following task:

Updates file metadata and/or content.

Arguments

request - No description provided.

fileId - The ID of the file to update.

pub fn watch(&self, request: Channel, file_id: &str) -> FileWatchCall<'a, C>

Create a builder to help you perform the following task:

Subscribe to changes on a file

Arguments

request - No description provided.

fileId - The ID for the file in question.