Struct google_drive2::About

pub struct About {
    pub language_code: Option<String>,
    pub quota_bytes_by_service: Option<Vec<AboutQuotaBytesByService>>,
    pub features: Option<Vec<AboutFeatures>>,
    pub quota_bytes_used: Option<String>,
    pub kind: Option<String>,
    pub additional_role_info: Option<Vec<AboutAdditionalRoleInfo>>,
    pub folder_color_palette: Option<Vec<String>>,
    pub user: Option<User>,
    pub quota_type: Option<String>,
    pub quota_bytes_used_aggregate: Option<String>,
    pub max_upload_sizes: Option<Vec<AboutMaxUploadSizes>>,
    pub permission_id: Option<String>,
    pub name: Option<String>,
    pub quota_bytes_total: Option<String>,
    pub remaining_change_ids: Option<String>,
    pub etag: Option<String>,
    pub import_formats: Option<Vec<AboutImportFormats>>,
    pub root_folder_id: Option<String>,
    pub largest_change_id: Option<String>,
    pub quota_bytes_used_in_trash: Option<String>,
    pub export_formats: Option<Vec<AboutExportFormats>>,
    pub domain_sharing_policy: Option<String>,
    pub self_link: Option<String>,
    pub is_current_app_installed: Option<bool>,
}

An item with user information and settings.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get about (response)

Fields

language_code

The user's language or locale code, as defined by BCP 47, with some extensions from Unicode's LDML format (http://www.unicode.org/reports/tr35/).

quota_bytes_by_service

The amount of storage quota used by different Google services.

features

List of additional features enabled on this account.

quota_bytes_used

The number of quota bytes used by Google Drive.

kind

This is always drive#about.

additional_role_info

Information about supported additional roles per file type. The most specific type takes precedence.

folder_color_palette

The palette of allowable folder colors as RGB hex strings.

user

The authenticated user.

quota_type

The type of the user's storage quota. Possible values are: - LIMITED - UNLIMITED

quota_bytes_used_aggregate

The number of quota bytes used by all Google apps (Drive, Picasa, etc.).

max_upload_sizes

List of max upload sizes for each file type. The most specific type takes precedence.

permission_id

The current user's ID as visible in the permissions collection.

name

The name of the current user.

quota_bytes_total

The total number of quota bytes.

remaining_change_ids

The number of remaining change ids.

etag

The ETag of the item.

import_formats

The allowable import formats.

root_folder_id

The id of the root folder.

largest_change_id

The largest change id.

quota_bytes_used_in_trash

The number of quota bytes used by trashed items.

export_formats

The allowable export formats.

domain_sharing_policy

The domain sharing policy for the current user.

self_link

A link back to this item.

is_current_app_installed

A boolean indicating whether the authenticated app is installed by the authenticated user.