Struct google_drive2::App

pub struct App {
    pub secondary_mime_types: Option<Vec<String>>,
    pub use_by_default: Option<bool>,
    pub installed: Option<bool>,
    pub product_url: Option<String>,
    pub supports_import: Option<bool>,
    pub supports_create: Option<bool>,
    pub long_description: Option<String>,
    pub short_description: Option<String>,
    pub id: Option<String>,
    pub has_drive_wide_scope: Option<bool>,
    pub object_type: Option<String>,
    pub kind: Option<String>,
    pub primary_file_extensions: Option<Vec<String>>,
    pub primary_mime_types: Option<Vec<String>>,
    pub create_url: Option<String>,
    pub icons: Option<Vec<AppIcons>>,
    pub supports_offline_create: Option<bool>,
    pub secondary_file_extensions: Option<Vec<String>>,
    pub name: Option<String>,
    pub authorized: Option<bool>,
    pub supports_multi_open: Option<bool>,
    pub create_in_folder_template: Option<String>,
    pub open_url_template: Option<String>,
    pub product_id: Option<String>,
}

The apps resource provides a list of the apps that a user has installed, with information about each app's supported MIME types, file extensions, and other details.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get apps (response)

list apps (none)

Fields
secondary_mime_types: Option<Vec<String>>

The list of secondary mime types.

use_by_default: Option<bool>

Whether the app is selected as the default handler for the types it supports.

installed: Option<bool>

Whether the app is installed.

product_url: Option<String>

A link to the product listing for this app.

supports_import: Option<bool>

Whether this app supports importing Google Docs.

supports_create: Option<bool>

Whether this app supports creating new objects.

long_description: Option<String>

A long description of the app.

short_description: Option<String>

A short description of the app.

id: Option<String>

The ID of the app.

has_drive_wide_scope: Option<bool>

Whether the app has drive-wide scope. An app with drive-wide scope can access all files in the user's drive.

object_type: Option<String>

The type of object this app creates (e.g. Chart). If empty, the app name should be used instead.

kind: Option<String>

This is always drive#app.

primary_file_extensions: Option<Vec<String>>

The list of primary file extensions.

primary_mime_types: Option<Vec<String>>

The list of primary mime types.

create_url: Option<String>

The url to create a new file with this app.

icons: Option<Vec<AppIcons>>

The various icons for the app.

supports_offline_create: Option<bool>

Whether this app supports creating new files when offline.

secondary_file_extensions: Option<Vec<String>>

The list of secondary file extensions.

name: Option<String>

The name of the app.

authorized: Option<bool>

Whether the app is authorized to access data on the user's Drive.

supports_multi_open: Option<bool>

Whether this app supports opening more than one file.

create_in_folder_template: Option<String>

The template url to create a new file with this app in a given folder. The template will contain {folderId} to be replaced by the folder to create the new file in.

open_url_template: Option<String>

The template url for opening files with this app. The template will contain {ids} and/or {exportIds} to be replaced by the actual file ids. See Open Files for the full documentation.

product_id: Option<String>

The ID of the product listing for this app.