Struct google_drive2::Comment

pub struct Comment {
    pub status: Option<String>,
    pub self_link: Option<String>,
    pub kind: Option<String>,
    pub author: Option<User>,
    pub deleted: Option<bool>,
    pub created_date: Option<String>,
    pub html_content: Option<String>,
    pub content: Option<String>,
    pub modified_date: Option<String>,
    pub context: Option<CommentContext>,
    pub replies: Option<Vec<CommentReply>>,
    pub comment_id: Option<String>,
    pub anchor: Option<String>,
    pub file_title: Option<String>,
    pub file_id: Option<String>,
}

A comment on a file in Google Drive.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update comments (request|response)

insert comments (request|response)

delete comments (none)

patch comments (request|response)

get comments (response)

list comments (none)

Fields

status

The status of this comment. Status can be changed by posting a reply to a comment with the desired status. - "open" - The comment is still open. - "resolved" - The comment has been resolved by one of its replies.

self_link

A link back to this comment.

kind

This is always drive#comment.

author

The user who wrote this comment.

deleted

Whether this comment has been deleted. If a comment has been deleted the content will be cleared and this will only represent a comment that once existed.

created_date

The date when this comment was first created.

html_content

HTML formatted content for this comment.

content

The plain text content used to create this comment. This is not HTML safe and should only be used as a starting point to make edits to a comment's content.

modified_date

The date when this comment or any of its replies were last modified.

context

The context of the file which is being commented on.

replies

Replies to this post.

comment_id

The ID of the comment.

anchor

A region of the document represented as a JSON string. See anchor documentation for details on how to define and interpret anchor properties.

file_title

The title of the file which this comment is addressing.

file_id

The file which this comment is addressing.