Struct google-drive2::CommentMethodsBuilder

pub struct CommentMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on comment resources. It is not used directly, but through the Drive hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-drive2" as drive2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use drive2::Drive;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Drive::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.comments();

Methods

impl<'a, C, NC, A> CommentMethodsBuilder<'a, C, NC, A>

fn delete(&self, file_id: &str, comment_id: &str) -> CommentDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes a comment.

fn get(&self, file_id: &str, comment_id: &str) -> CommentGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a comment by ID.

fn insert(&self, request: &Comment, file_id: &str) -> CommentInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a new comment on the given file.

fn patch(&self, request: &Comment, file_id: &str, comment_id: &str) -> CommentPatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing comment. This method supports patch semantics.

fn update(&self, request: &Comment, file_id: &str, comment_id: &str) -> CommentUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates an existing comment.

fn list(&self, file_id: &str) -> CommentListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists a file's comments.