Struct google-drive2::Drive

pub struct Drive<C, NC, A> {
    // some fields omitted
}

Central instance to access all Drive related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-drive2" as drive2;
use drive2::File;
use drive2::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use drive2::Drive;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Drive::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: File = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.files().patch(&req, "fileId")
             .use_content_as_indexable_text(true)
             .update_viewed_date(true)
             .timed_text_track_name("duo")
             .timed_text_language("aliquyam")
             .set_modified_date(true)
             .remove_parents("Lorem")
             .pinned(false)
             .ocr_language("erat")
             .ocr(false)
             .new_revision(true)
             .convert(true)
             .add_parents("elitr")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Methods

impl<'a, C, NC, A> Drive<C, NC, A> where NC: NetworkConnector, C: BorrowMut<Client<NC>>, A: GetToken

fn new(client: C, authenticator: A) -> Drive<C, NC, A>

fn about(&'a self) -> AboutMethodsBuilder<'a, C, NC, A>

fn apps(&'a self) -> AppMethodsBuilder<'a, C, NC, A>

fn changes(&'a self) -> ChangeMethodsBuilder<'a, C, NC, A>

fn channels(&'a self) -> ChannelMethodsBuilder<'a, C, NC, A>

fn children(&'a self) -> ChildrenMethodsBuilder<'a, C, NC, A>

fn comments(&'a self) -> CommentMethodsBuilder<'a, C, NC, A>

fn files(&'a self) -> FileMethodsBuilder<'a, C, NC, A>

fn parents(&'a self) -> ParentMethodsBuilder<'a, C, NC, A>

fn permissions(&'a self) -> PermissionMethodsBuilder<'a, C, NC, A>

fn properties(&'a self) -> PropertyMethodsBuilder<'a, C, NC, A>

fn realtime(&'a self) -> RealtimeMethodsBuilder<'a, C, NC, A>

fn replies(&'a self) -> ReplyMethodsBuilder<'a, C, NC, A>

fn revisions(&'a self) -> RevisionMethodsBuilder<'a, C, NC, A>

fn user_agent(&mut self, agent_name: String) -> String

Set the user-agent header field to use in all requests to the server. It defaults to google-api-rust-client/0.0.1.

Returns the previously set user-agent.