Struct google_drive2::FileMethods

pub struct FileMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on file resources. It is not used directly, but through the Drive hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_drive2 as drive2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use drive2::Drive;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Drive::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `copy(...)`, `delete(...)`, `empty_trash(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)`, `touch(...)`, `trash(...)`, `untrash(...)`, `update(...)` and `watch(...)`
// to build up your call.
let rb = hub.files();

Methods

impl<'a, C, A> FileMethods<'a, C, A>

fn watch(&self, request: Channel, file_id: &str) -> FileWatchCall<'a, C, A>

Create a builder to help you perform the following task:

Subscribe to changes on a file

Arguments

request - No description provided.

fileId - The ID for the file in question.

fn insert(&self, request: File) -> FileInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Insert a new file.

Arguments

request - No description provided.

fn untrash(&self, file_id: &str) -> FileUntrashCall<'a, C, A>

Create a builder to help you perform the following task:

Restores a file from the trash.

Arguments

fileId - The ID of the file to untrash.

fn copy(&self, request: File, file_id: &str) -> FileCopyCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a copy of the specified file.

Arguments

request - No description provided.

fileId - The ID of the file to copy.

fn trash(&self, file_id: &str) -> FileTrashCall<'a, C, A>

Create a builder to help you perform the following task:

Moves a file to the trash.

Arguments

fileId - The ID of the file to trash.

fn empty_trash(&self) -> FileEmptyTrashCall<'a, C, A>

Create a builder to help you perform the following task:

Permanently deletes all of the user's trashed files.

fn list(&self) -> FileListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the user's files.

fn touch(&self, file_id: &str) -> FileTouchCall<'a, C, A>

Create a builder to help you perform the following task:

Set the file's updated time to the current server time.

Arguments

fileId - The ID of the file to update.

fn update(&self, request: File, file_id: &str) -> FileUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates file metadata and/or content.

Arguments

request - No description provided.

fileId - The ID of the file to update.

fn delete(&self, file_id: &str) -> FileDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Permanently deletes a file by ID. Skips the trash. The currently authenticated user must own the file.

Arguments

fileId - The ID of the file to delete.

fn patch(&self, request: File, file_id: &str) -> FilePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates file metadata and/or content. This method supports patch semantics.

Arguments

request - No description provided.

fileId - The ID of the file to update.

fn get(&self, file_id: &str) -> FileGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a file's metadata by ID.

Arguments

fileId - The ID for the file in question.