Struct google_drive2::Permission

pub struct Permission {
    pub with_link: Option<bool>,
    pub domain: Option<String>,
    pub name: Option<String>,
    pub kind: Option<String>,
    pub value: Option<String>,
    pub additional_roles: Option<Vec<String>>,
    pub auth_key: Option<String>,
    pub etag: Option<String>,
    pub email_address: Option<String>,
    pub photo_link: Option<String>,
    pub role: Option<String>,
    pub type_: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
}

A permission for a file.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch permissions (request|response)

list permissions (none)

insert permissions (request|response)

get permissions (response)

get id for email permissions (none)

delete permissions (none)

update permissions (request|response)

Fields

with_link

Whether the link is required for this permission.

domain

The domain name of the entity this permission refers to. This is an output-only field which is present when the permission type is user, group or domain.

name

The name for this permission.

kind

This is always drive#permission.

value

The email address or domain name for the entity. This is used during inserts and is not populated in responses. When making a drive.permissions.insert request, exactly one of the id or value fields must be specified.

additional_roles

Additional roles for this user. Only commenter is currently allowed.

auth_key

The authkey parameter required for this permission.

etag

The ETag of the permission.

email_address

The email address of the user or group this permission refers to. This is an output-only field which is present when the permission type is user or group.

photo_link

A link to the profile photo, if available.

role

The primary role for this user. Allowed values are: - owner - reader - writer

type_

The account type. Allowed values are: - user - group - domain - anyone

id

The ID of the user this permission refers to, and identical to the permissionId in the About and Files resources. When making a drive.permissions.insert request, exactly one of the id or value fields must be specified.

self_link

A link back to this permission.