Struct google_drive2::PropertyMethods

pub struct PropertyMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on property resources. It is not used directly, but through the Drive hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_drive2 as drive2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use drive2::Drive;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Drive::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.properties();

Methods

impl<'a, C, A> PropertyMethods<'a, C, A>

fn patch(&self, request: Property, file_id: &str, property_key: &str) -> PropertyPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a property. This method supports patch semantics.

Arguments

request - No description provided.

fileId - The ID of the file.

propertyKey - The key of the property.

fn delete(&self, file_id: &str, property_key: &str) -> PropertyDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a property.

Arguments

fileId - The ID of the file.

propertyKey - The key of the property.

fn insert(&self, request: Property, file_id: &str) -> PropertyInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Adds a property to a file.

Arguments

request - No description provided.

fileId - The ID of the file.

fn list(&self, file_id: &str) -> PropertyListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists a file's properties.

Arguments

fileId - The ID of the file.

fn update(&self, request: Property, file_id: &str, property_key: &str) -> PropertyUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a property.

Arguments

request - No description provided.

fileId - The ID of the file.

propertyKey - The key of the property.

fn get(&self, file_id: &str, property_key: &str) -> PropertyGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a property by its key.

Arguments

fileId - The ID of the file.

propertyKey - The key of the property.