Struct google_drive2::Revision

pub struct Revision {
    pub mime_type: Option<String>,
    pub published_link: Option<String>,
    pub published_outside_domain: Option<bool>,
    pub publish_auto: Option<bool>,
    pub file_size: Option<String>,
    pub last_modifying_user_name: Option<String>,
    pub id: Option<String>,
    pub md5_checksum: Option<String>,
    pub download_url: Option<String>,
    pub kind: Option<String>,
    pub last_modifying_user: Option<User>,
    pub pinned: Option<bool>,
    pub etag: Option<String>,
    pub original_filename: Option<String>,
    pub export_links: Option<HashMap<String, String>>,
    pub published: Option<bool>,
    pub self_link: Option<String>,
    pub modified_date: Option<String>,
}

A revision of a file.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete revisions (none)

update revisions (request|response)

patch revisions (request|response)

list revisions (none)

get revisions (response)

Fields

mime_type

The MIME type of the revision.

published_link

A link to the published revision.

published_outside_domain

Whether this revision is published outside the domain. This is only populated and can only be modified for Google Docs.

publish_auto

Whether subsequent revisions will be automatically republished. This is only populated and can only be modified for Google Docs.

file_size

The size of the revision in bytes. This will only be populated on files with content stored in Drive.

last_modifying_user_name

Name of the last user to modify this revision.

id

The ID of the revision.

md5_checksum

An MD5 checksum for the content of this revision. This will only be populated on files with content stored in Drive.

download_url

Short term download URL for the file. This will only be populated on files with content stored in Drive.

kind

This is always drive#revision.

last_modifying_user

The last user to modify this revision.

pinned

Whether this revision is pinned to prevent automatic purging. This will only be populated and can only be modified on files with content stored in Drive which are not Google Docs. Revisions can also be pinned when they are created through the drive.files.insert/update/copy by using the pinned query parameter.

etag

The ETag of the revision.

original_filename

The original filename when this revision was created. This will only be populated on files with content stored in Drive.

export_links

Links for exporting Google Docs to specific formats.

published

Whether this revision is published. This is only populated and can only be modified for Google Docs.

self_link

A link back to this revision.

modified_date

Last time this revision was modified (formatted RFC 3339 timestamp).