Struct google_drive2::TeamDriveCapabilities

pub struct TeamDriveCapabilities {
    pub can_read_revisions: Option<bool>,
    pub can_copy: Option<bool>,
    pub can_change_copy_requires_writer_permission_restriction: Option<bool>,
    pub can_trash_children: Option<bool>,
    pub can_change_domain_users_only_restriction: Option<bool>,
    pub can_delete_team_drive: Option<bool>,
    pub can_rename_team_drive: Option<bool>,
    pub can_comment: Option<bool>,
    pub can_list_children: Option<bool>,
    pub can_rename: Option<bool>,
    pub can_delete_children: Option<bool>,
    pub can_add_children: Option<bool>,
    pub can_share: Option<bool>,
    pub can_manage_members: Option<bool>,
    pub can_download: Option<bool>,
    pub can_change_team_members_only_restriction: Option<bool>,
    pub can_change_team_drive_background: Option<bool>,
    pub can_remove_children: Option<bool>,
    pub can_edit: Option<bool>,
}

Capabilities the current user has on this Team Drive.

This type is not used in any activity, and only used as part of another schema.

Fields
can_read_revisions: Option<bool>

Whether the current user can read the revisions resource of files in this Team Drive.

can_copy: Option<bool>

Whether the current user can copy files in this Team Drive.

can_change_copy_requires_writer_permission_restriction: Option<bool>

Whether the current user can change the copyRequiresWriterPermission restriction of this Team Drive.

can_trash_children: Option<bool>

Whether the current user can trash children from folders in this Team Drive.

can_change_domain_users_only_restriction: Option<bool>

Whether the current user can change the domainUsersOnly restriction of this Team Drive.

can_delete_team_drive: Option<bool>

Whether the current user can delete this Team Drive. Attempting to delete the Team Drive may still fail if there are untrashed items inside the Team Drive.

can_rename_team_drive: Option<bool>

Whether the current user can rename this Team Drive.

can_comment: Option<bool>

Whether the current user can comment on files in this Team Drive.

can_list_children: Option<bool>

Whether the current user can list the children of folders in this Team Drive.

can_rename: Option<bool>

Whether the current user can rename files or folders in this Team Drive.

can_delete_children: Option<bool>

Whether the current user can delete children from folders in this Team Drive.

can_add_children: Option<bool>

Whether the current user can add children to folders in this Team Drive.

can_share: Option<bool>

Whether the current user can share files or folders in this Team Drive.

can_manage_members: Option<bool>

Whether the current user can add members to this Team Drive or remove them or change their role.

can_download: Option<bool>

Whether the current user can download files in this Team Drive.

can_change_team_members_only_restriction: Option<bool>

Whether the current user can change the teamMembersOnly restriction of this Team Drive.

can_change_team_drive_background: Option<bool>

Whether the current user can change the background of this Team Drive.

can_remove_children: Option<bool>

Deprecated - use canDeleteChildren or canTrashChildren instead.

can_edit: Option<bool>

Whether the current user can edit files in this Team Drive