Crate google_drive2

This documentation was generated from drive crate version 1.0.9+20190620, where 20190620 is the exact revision of the drive:v2 schema built by the mako code generator v1.0.9.

Everything else about the drive v2 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

about

get

apps

get and list

changes

get, get start page token, list and watch

channels

stop

children

delete, get, insert and list

comments

delete, get, insert, list, patch and update

drives

delete, get, hide, insert, list, unhide and update

files

copy, delete, empty trash, export, generate ids, get, insert, list, patch, touch, trash, untrash, update and watch

parents

delete, get, insert and list

permissions

delete, get, get id for email, insert, list, patch and update

properties

delete, get, insert, list, patch and update

realtime

get and update

replies

delete, get, insert, list, patch and update

revisions

delete, get, list, patch and update

teamdrives

delete, get, insert, list and update

Upload supported by ...

update files

insert files

update realtime

Download supported by ...

watch files

get realtime

get files

export files

Subscription supported by ...

watch files

insert files

get files

watch changes

list changes

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.files().watch(...).doit()
let r = hub.files().empty_trash(...).doit()
let r = hub.files().generate_ids(...).doit()
let r = hub.files().copy(...).doit()
let r = hub.files().list(...).doit()
let r = hub.files().delete(...).doit()
let r = hub.files().patch(...).doit()
let r = hub.files().update(...).doit()
let r = hub.files().insert(...).doit()
let r = hub.files().untrash(...).doit()
let r = hub.files().trash(...).doit()
let r = hub.files().touch(...).doit()
let r = hub.files().get(...).doit()
let r = hub.files().export(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-drive2 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_drive2 as drive2;
use drive2::File;
use drive2::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use drive2::DriveHub;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = DriveHub::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = File::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.files().patch(req, "fileId")
             .use_content_as_indexable_text(false)
             .update_viewed_date(true)
             .timed_text_track_name("ea")
             .timed_text_language("no")
             .supports_team_drives(true)
             .supports_all_drives(true)
             .set_modified_date(true)
             .remove_parents("et")
             .pinned(true)
             .ocr_language("ipsum")
             .ocr(true)
             .new_revision(true)
             .modified_date_behavior("duo")
             .convert(true)
             .add_parents("sea")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

About
An item with user information and settings.

AboutAdditionalRoleInfo
Information about supported additional roles per file type. The most specific type takes precedence.

AboutAdditionalRoleInfoRoleSets
The supported additional roles per primary role.

AboutDriveThemes
A list of themes that are supported for shared drives.

AboutExportFormats
The allowable export formats.

AboutFeatures
List of additional features enabled on this account.

AboutGetCall
Gets the information about the current user along with Drive API settings

AboutImportFormats
The allowable import formats.

AboutMaxUploadSizes
List of max upload sizes for each file type. The most specific type takes precedence.

AboutMethods
A builder providing access to all methods supported on about resources. It is not used directly, but through the DriveHub hub.

AboutQuotaBytesByService
The amount of storage quota used by different Google services.

AboutTeamDriveThemes
Deprecated - use driveThemes instead.

App
The apps resource provides a list of the apps that a user has installed, with information about each app's supported MIME types, file extensions, and other details.

AppGetCall
Gets a specific app.

AppIcons
The various icons for the app.

AppList
A list of third-party applications which the user has installed or given access to Google Drive.

AppListCall
Lists a user's installed apps.

AppMethods
A builder providing access to all methods supported on app resources. It is not used directly, but through the DriveHub hub.

Change
Representation of a change to a file or shared drive.

ChangeGetCall
Deprecated - Use changes.getStartPageToken and changes.list to retrieve recent changes.

ChangeGetStartPageTokenCall
Gets the starting pageToken for listing future changes.

ChangeList
A list of changes for a user.

ChangeListCall
Lists the changes for a user or shared drive.

ChangeMethods
A builder providing access to all methods supported on change resources. It is not used directly, but through the DriveHub hub.

ChangeWatchCall
Subscribe to changes for a user.

Channel
An notification channel used to watch for resource changes.

ChannelMethods
A builder providing access to all methods supported on channel resources. It is not used directly, but through the DriveHub hub.

ChannelStopCall
Stop watching resources through this channel

ChildList
A list of children of a file.

ChildReference
A reference to a folder's child.

ChildrenDeleteCall
Removes a child from a folder.

ChildrenGetCall
Gets a specific child reference.

ChildrenInsertCall
Inserts a file into a folder.

ChildrenListCall
Lists a folder's children.

ChildrenMethods
A builder providing access to all methods supported on children resources. It is not used directly, but through the DriveHub hub.

Chunk

Comment
A comment on a file in Google Drive.

CommentContext
The context of the file which is being commented on.

CommentDeleteCall
Deletes a comment.

CommentGetCall
Gets a comment by ID.

CommentInsertCall
Creates a new comment on the given file.

CommentList
A list of comments on a file in Google Drive.

CommentListCall
Lists a file's comments.

CommentMethods
A builder providing access to all methods supported on comment resources. It is not used directly, but through the DriveHub hub.

CommentPatchCall
Updates an existing comment. This method supports patch semantics.

CommentReply
A comment on a file in Google Drive.

CommentReplyList
A list of replies to a comment on a file in Google Drive.

CommentUpdateCall
Updates an existing comment.

ContentRange
Implements the Content-Range header, for serialization only

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Drive
Representation of a shared drive.

DriveBackgroundImageFile
An image file and cropping parameters from which a background image for this shared drive is set. This is a write only field; it can only be set on drive.drives.update requests that don't set themeId. When specified, all fields of the backgroundImageFile must be set.

DriveCapabilities
Capabilities the current user has on this shared drive.

DriveDeleteCall
Permanently deletes a shared drive for which the user is an organizer. The shared drive cannot contain any untrashed items.

DriveGetCall
Gets a shared drive's metadata by ID.

DriveHideCall
Hides a shared drive from the default view.

DriveHub
Central instance to access all DriveHub related resource activities

DriveInsertCall
Creates a new shared drive.

DriveList
A list of shared drives.

DriveListCall
Lists the user's shared drives.

DriveMethods
A builder providing access to all methods supported on drive resources. It is not used directly, but through the DriveHub hub.

DriveRestrictions
A set of restrictions that apply to this shared drive or items inside this shared drive.

DriveUnhideCall
Restores a shared drive to the default view.

DriveUpdateCall
Updates the metadata for a shared drive.

DummyNetworkStream

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

File
The metadata for a file.

FileCapabilities
Capabilities the current user has on this file. Each capability corresponds to a fine-grained action that a user may take.

FileCopyCall
Creates a copy of the specified file.

FileDeleteCall
Permanently deletes a file by ID. Skips the trash. The currently authenticated user must own the file or be an organizer on the parent for shared drive files.

FileEmptyTrashCall
Permanently deletes all of the user's trashed files.

FileExportCall
Exports a Google Doc to the requested MIME type and returns the exported content. Please note that the exported content is limited to 10MB.

FileGenerateIdCall
Generates a set of file IDs which can be provided in insert or copy requests.

FileGetCall
Gets a file's metadata by ID.

FileImageMediaMetadata
Metadata about image media. This will only be present for image types, and its contents will depend on what can be parsed from the image content.

FileImageMediaMetadataLocation
Geographic location information stored in the image.

FileIndexableText
Indexable text attributes for the file (can only be written)

FileInsertCall
Insert a new file.

FileLabels
A group of labels for the file.

FileList
A list of files.

FileListCall
Lists the user's files.

FileMethods
A builder providing access to all methods supported on file resources. It is not used directly, but through the DriveHub hub.

FilePatchCall
Updates file metadata and/or content. This method supports patch semantics.

FileThumbnail
A thumbnail for the file. This will only be used if a standard thumbnail cannot be generated.

FileTouchCall
Set the file's updated time to the current server time.

FileTrashCall
Moves a file to the trash. The currently authenticated user must own the file or be at least a fileOrganizer on the parent for shared drive files.

FileUntrashCall
Restores a file from the trash.

FileUpdateCall
Updates file metadata and/or content.

FileVideoMediaMetadata
Metadata about video media. This will only be present for video types.

FileWatchCall
Subscribe to changes on a file

GeneratedIds
A list of generated IDs which can be provided in insert requests

JsonServerError
A utility type which can decode a server response that indicates error

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

ParentDeleteCall
Removes a parent from a file.

ParentGetCall
Gets a specific parent reference.

ParentInsertCall
Adds a parent folder for a file.

ParentList
A list of a file's parents.

ParentListCall
Lists a file's parents.

ParentMethods
A builder providing access to all methods supported on parent resources. It is not used directly, but through the DriveHub hub.

ParentReference
A reference to a file's parent.

Permission
A permission for a file.

PermissionDeleteCall
Deletes a permission from a file or shared drive.

PermissionGetCall
Gets a permission by ID.

PermissionGetIdForEmailCall
Returns the permission ID for an email address.

PermissionId
An ID for a user or group as seen in Permission items.

PermissionInsertCall
Inserts a permission for a file or shared drive.

PermissionList
A list of permissions associated with a file.

PermissionListCall
Lists a file's or shared drive's permissions.

PermissionMethods
A builder providing access to all methods supported on permission resources. It is not used directly, but through the DriveHub hub.

PermissionPatchCall
Updates a permission using patch semantics.

PermissionPermissionDetails
Details of whether the permissions on this shared drive item are inherited or directly on this item. This is an output-only field which is present only for shared drive items.

PermissionTeamDrivePermissionDetails
Deprecated - use permissionDetails instead.

PermissionUpdateCall
Updates a permission.

Property
A key-value pair attached to a file that is either public or private to an application. The following limits apply to file properties:

PropertyDeleteCall
Deletes a property.

PropertyGetCall
Gets a property by its key.

PropertyInsertCall
Adds a property to a file, or updates it if it already exists.

PropertyList
A collection of properties, key-value pairs that are either public or private to an application.

PropertyListCall
Lists a file's properties.

PropertyMethods
A builder providing access to all methods supported on property resources. It is not used directly, but through the DriveHub hub.

PropertyPatchCall
Updates a property.

PropertyUpdateCall
Updates a property.

RangeResponseHeader

RealtimeGetCall
Exports the contents of the Realtime API data model associated with this file as JSON.

RealtimeMethods
A builder providing access to all methods supported on realtime resources. It is not used directly, but through the DriveHub hub.

RealtimeUpdateCall
Overwrites the Realtime API data model associated with this file with the provided JSON data model.

ReplyDeleteCall
Deletes a reply.

ReplyGetCall
Gets a reply.

ReplyInsertCall
Creates a new reply to the given comment.

ReplyListCall
Lists all of the replies to a comment.

ReplyMethods
A builder providing access to all methods supported on reply resources. It is not used directly, but through the DriveHub hub.

ReplyPatchCall
Updates an existing reply. This method supports patch semantics.

ReplyUpdateCall
Updates an existing reply.

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

Revision
A revision of a file.

RevisionDeleteCall
Permanently deletes a file version. You can only delete revisions for files with binary content, like images or videos. Revisions for other files, like Google Docs or Sheets, and the last remaining file version can't be deleted.

RevisionGetCall
Gets a specific revision.

RevisionList
A list of revisions of a file.

RevisionListCall
Lists a file's revisions.

RevisionMethods
A builder providing access to all methods supported on revision resources. It is not used directly, but through the DriveHub hub.

RevisionPatchCall
Updates a revision. This method supports patch semantics.

RevisionUpdateCall
Updates a revision.

ServerError

ServerMessage

StartPageToken
There is no detailed description.

TeamDrive
Deprecated: use the drive collection instead.

TeamDriveBackgroundImageFile
An image file and cropping parameters from which a background image for this Team Drive is set. This is a write only field; it can only be set on drive.teamdrives.update requests that don't set themeId. When specified, all fields of the backgroundImageFile must be set.

TeamDriveCapabilities
Capabilities the current user has on this Team Drive.

TeamDriveList
A list of Team Drives.

TeamDriveRestrictions
A set of restrictions that apply to this Team Drive or items inside this Team Drive.

TeamdriveDeleteCall
Deprecated use drives.delete instead.

TeamdriveGetCall
Deprecated use drives.get instead.

TeamdriveInsertCall
Deprecated use drives.insert instead.

TeamdriveListCall
Deprecated use drives.list instead.

TeamdriveMethods
A builder providing access to all methods supported on teamdrive resources. It is not used directly, but through the DriveHub hub.

TeamdriveUpdateCall
Deprecated use drives.update instead.

User
Information about a Drive user.

UserPicture
The user's profile picture.

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.