Crate google_drive2

This documentation was generated from drive crate version 0.1.11+20160126, where 20160126 is the exact revision of the drive:v2 schema built by the mako code generator v0.1.11.

Everything else about the drive v2 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

about

get

apps

get and list

changes

get, list and watch

channels

stop

children

delete, get, insert and list

comments

delete, get, insert, list, patch and update

files

copy, delete, empty trash, generate ids, get, insert, list, patch, touch, trash, untrash, update and watch

parents

delete, get, insert and list

permissions

delete, get, get id for email, insert, list, patch and update

properties

delete, get, insert, list, patch and update

realtime

get and update

replies

delete, get, insert, list, patch and update

revisions

delete, get, list, patch and update

Upload supported by ...

insert files

update realtime

update files

Download supported by ...

watch files

get realtime

get files

Subscription supported by ...

watch files

watch changes

insert files

list changes

get files

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.files().watch(...).doit()
let r = hub.files().empty_trash(...).doit()
let r = hub.files().generate_ids(...).doit()
let r = hub.files().copy(...).doit()
let r = hub.files().list(...).doit()
let r = hub.files().delete(...).doit()
let r = hub.files().patch(...).doit()
let r = hub.files().update(...).doit()
let r = hub.files().insert(...).doit()
let r = hub.files().untrash(...).doit()
let r = hub.files().trash(...).doit()
let r = hub.files().touch(...).doit()
let r = hub.files().get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-drive2 = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_drive2 as drive2;
use drive2::File;
use drive2::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use drive2::Drive;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Drive::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = File::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.files().patch(req, "fileId")
             .use_content_as_indexable_text(true)
             .update_viewed_date(true)
             .timed_text_track_name("Stet")
             .timed_text_language("sed")
             .set_modified_date(false)
             .remove_parents("sanctus")
             .pinned(true)
             .ocr_language("Lorem")
             .ocr(false)
             .new_revision(true)
             .modified_date_behavior("eirmod")
             .convert(true)
             .add_parents("gubergren")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

About
An item with user information and settings.

AboutAdditionalRoleInfo
Information about supported additional roles per file type. The most specific type takes precedence.

AboutAdditionalRoleInfoRoleSets
The supported additional roles per primary role.

AboutExportFormats
The allowable export formats.

AboutFeatures
List of additional features enabled on this account.

AboutGetCall
Gets the information about the current user along with Drive API settings

AboutImportFormats
The allowable import formats.

AboutMaxUploadSizes
List of max upload sizes for each file type. The most specific type takes precedence.

AboutMethods
A builder providing access to all methods supported on about resources. It is not used directly, but through the Drive hub.

AboutQuotaBytesByService
The amount of storage quota used by different Google services.

App
The apps resource provides a list of the apps that a user has installed, with information about each app's supported MIME types, file extensions, and other details.

AppGetCall
Gets a specific app.

AppIcons
The various icons for the app.

AppList
A list of third-party applications which the user has installed or given access to Google Drive.

AppListCall
Lists a user's installed apps.

AppMethods
A builder providing access to all methods supported on app resources. It is not used directly, but through the Drive hub.

Change
Representation of a change to a file.

ChangeGetCall
Gets a specific change.

ChangeList
A list of changes for a user.

ChangeListCall
Lists the changes for a user.

ChangeMethods
A builder providing access to all methods supported on change resources. It is not used directly, but through the Drive hub.

ChangeWatchCall
Subscribe to changes for a user.

Channel
An notification channel used to watch for resource changes.

ChannelMethods
A builder providing access to all methods supported on channel resources. It is not used directly, but through the Drive hub.

ChannelStopCall
Stop watching resources through this channel

ChildList
A list of children of a file.

ChildReference
A reference to a folder's child.

ChildrenDeleteCall
Removes a child from a folder.

ChildrenGetCall
Gets a specific child reference.

ChildrenInsertCall
Inserts a file into a folder.

ChildrenListCall
Lists a folder's children.

ChildrenMethods
A builder providing access to all methods supported on children resources. It is not used directly, but through the Drive hub.

Comment
A comment on a file in Google Drive.

CommentContext
The context of the file which is being commented on.

CommentDeleteCall
Deletes a comment.

CommentGetCall
Gets a comment by ID.

CommentInsertCall
Creates a new comment on the given file.

CommentList
A list of comments on a file in Google Drive.

CommentListCall
Lists a file's comments.

CommentMethods
A builder providing access to all methods supported on comment resources. It is not used directly, but through the Drive hub.

CommentPatchCall
Updates an existing comment. This method supports patch semantics.

CommentReply
A comment on a file in Google Drive.

CommentReplyList
A list of replies to a comment on a file in Google Drive.

CommentUpdateCall
Updates an existing comment.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Drive
Central instance to access all Drive related resource activities

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

File
The metadata for a file.

FileCopyCall
Creates a copy of the specified file.

FileDeleteCall
Permanently deletes a file by ID. Skips the trash. The currently authenticated user must own the file.

FileEmptyTrashCall
Permanently deletes all of the user's trashed files.

FileGenerateIdCall
Generates a set of file IDs which can be provided in insert requests.

FileGetCall
Gets a file's metadata by ID.

FileImageMediaMetadata
Metadata about image media. This will only be present for image types, and its contents will depend on what can be parsed from the image content.

FileImageMediaMetadataLocation
Geographic location information stored in the image.

FileIndexableText
Indexable text attributes for the file (can only be written)

FileInsertCall
Insert a new file.

FileLabels
A group of labels for the file.

FileList
A list of files.

FileListCall
Lists the user's files.

FileMethods
A builder providing access to all methods supported on file resources. It is not used directly, but through the Drive hub.

FilePatchCall
Updates file metadata and/or content. This method supports patch semantics.

FileThumbnail
Thumbnail for the file. Only accepted on upload and for files that are not already thumbnailed by Google.

FileTouchCall
Set the file's updated time to the current server time.

FileTrashCall
Moves a file to the trash. The currently authenticated user must own the file.

FileUntrashCall
Restores a file from the trash.

FileUpdateCall
Updates file metadata and/or content.

FileVideoMediaMetadata
Metadata about video media. This will only be present for video types.

FileWatchCall
Subscribe to changes on a file

GeneratedIds
A list of generated IDs which can be provided in insert requests

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

ParentDeleteCall
Removes a parent from a file.

ParentGetCall
Gets a specific parent reference.

ParentInsertCall
Adds a parent folder for a file.

ParentList
A list of a file's parents.

ParentListCall
Lists a file's parents.

ParentMethods
A builder providing access to all methods supported on parent resources. It is not used directly, but through the Drive hub.

ParentReference
A reference to a file's parent.

Permission
A permission for a file.

PermissionDeleteCall
Deletes a permission from a file.

PermissionGetCall
Gets a permission by ID.

PermissionGetIdForEmailCall
Returns the permission ID for an email address.

PermissionId
An ID for a user or group as seen in Permission items.

PermissionInsertCall
Inserts a permission for a file.

PermissionList
A list of permissions associated with a file.

PermissionListCall
Lists a file's permissions.

PermissionMethods
A builder providing access to all methods supported on permission resources. It is not used directly, but through the Drive hub.

PermissionPatchCall
Updates a permission using patch semantics.

PermissionUpdateCall
Updates a permission.

Property
A key-value pair attached to a file that is either public or private to an application. The following limits apply to file properties:
- Maximum of 100 properties total per file - Maximum of 30 private properties per app - Maximum of 30 public properties - Maximum of 124 bytes size limit on (key + value) string in UTF-8 encoding for a single property.

PropertyDeleteCall
Deletes a property.

PropertyGetCall
Gets a property by its key.

PropertyInsertCall
Adds a property to a file, or updates it if it already exists.

PropertyList
A collection of properties, key-value pairs that are either public or private to an application.

PropertyListCall
Lists a file's properties.

PropertyMethods
A builder providing access to all methods supported on property resources. It is not used directly, but through the Drive hub.

PropertyPatchCall
Updates a property, or adds it if it doesn't exist. This method supports patch semantics.

PropertyUpdateCall
Updates a property, or adds it if it doesn't exist.

RealtimeGetCall
Exports the contents of the Realtime API data model associated with this file as JSON.

RealtimeMethods
A builder providing access to all methods supported on realtime resources. It is not used directly, but through the Drive hub.

RealtimeUpdateCall
Overwrites the Realtime API data model associated with this file with the provided JSON data model.

ReplyDeleteCall
Deletes a reply.

ReplyGetCall
Gets a reply.

ReplyInsertCall
Creates a new reply to the given comment.

ReplyListCall
Lists all of the replies to a comment.

ReplyMethods
A builder providing access to all methods supported on reply resources. It is not used directly, but through the Drive hub.

ReplyPatchCall
Updates an existing reply. This method supports patch semantics.

ReplyUpdateCall
Updates an existing reply.

Revision
A revision of a file.

RevisionDeleteCall
Removes a revision.

RevisionGetCall
Gets a specific revision.

RevisionList
A list of revisions of a file.

RevisionListCall
Lists a file's revisions.

RevisionMethods
A builder providing access to all methods supported on revision resources. It is not used directly, but through the Drive hub.

RevisionPatchCall
Updates a revision. This method supports patch semantics.

RevisionUpdateCall
Updates a revision.

User
Information about a Drive user.

UserPicture
The user's profile picture.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.