Struct google_drive2::CommentReply

pub struct CommentReply {
    pub kind: Option<String>,
    pub author: Option<User>,
    pub deleted: Option<bool>,
    pub html_content: Option<String>,
    pub content: Option<String>,
    pub verb: Option<String>,
    pub reply_id: Option<String>,
    pub modified_date: Option<String>,
    pub created_date: Option<String>,
}

A JSON representation of a reply to a comment on a file in Google Drive.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch replies (request|response)

insert replies (request|response)

get replies (response)

update replies (request|response)

Fields

kind

This is always drive#commentReply.

author

The user who wrote this reply.

deleted

Whether this reply has been deleted. If a reply has been deleted the content will be cleared and this will only represent a reply that once existed.

html_content

HTML formatted content for this reply.

content

The plain text content used to create this reply. This is not HTML safe and should only be used as a starting point to make edits to a reply's content. This field is required on inserts if no verb is specified (resolve/reopen).

verb

The action this reply performed to the parent comment. When creating a new reply this is the action to be perform to the parent comment. Possible values are: - "resolve" - To resolve a comment. - "reopen" - To reopen (un-resolve) a comment.

reply_id

The ID of the reply.

modified_date

The date when this reply was last modified.

created_date

The date when this reply was first created.