Struct google_drive2::File

pub struct File {
    pub mime_type: Option<String>,
    pub last_viewed_by_me_date: Option<String>,
    pub app_data_contents: Option<bool>,
    pub thumbnail_link: Option<String>,
    pub labels: Option<FileLabels>,
    pub indexable_text: Option<FileIndexableText>,
    pub explicitly_trashed: Option<bool>,
    pub is_app_authorized: Option<bool>,
    pub last_modifying_user_name: Option<String>,
    pub writers_can_share: Option<bool>,
    pub owned_by_me: Option<bool>,
    pub shared_with_me_date: Option<String>,
    pub sharing_user: Option<User>,
    pub video_media_metadata: Option<FileVideoMediaMetadata>,
    pub last_modifying_user: Option<User>,
    pub copyable: Option<bool>,
    pub folder_color_rgb: Option<String>,
    pub owner_names: Option<Vec<String>>,
    pub kind: Option<String>,
    pub id: Option<String>,
    pub web_view_link: Option<String>,
    pub version: Option<String>,
    pub parents: Option<Vec<ParentReference>>,
    pub export_links: Option<HashMap<String, String>>,
    pub shared: Option<bool>,
    pub icon_link: Option<String>,
    pub full_file_extension: Option<String>,
    pub open_with_links: Option<HashMap<String, String>>,
    pub can_read_revisions: Option<bool>,
    pub default_open_with_link: Option<String>,
    pub description: Option<String>,
    pub web_content_link: Option<String>,
    pub editable: Option<bool>,
    pub marked_viewed_by_me_date: Option<String>,
    pub embed_link: Option<String>,
    pub can_comment: Option<bool>,
    pub file_extension: Option<String>,
    pub modified_date: Option<String>,
    pub created_date: Option<String>,
    pub properties: Option<Vec<Property>>,
    pub md5_checksum: Option<String>,
    pub permissions: Option<Vec<Permission>>,
    pub image_media_metadata: Option<FileImageMediaMetadata>,
    pub owners: Option<Vec<User>>,
    pub etag: Option<String>,
    pub alternate_link: Option<String>,
    pub title: Option<String>,
    pub modified_by_me_date: Option<String>,
    pub thumbnail: Option<FileThumbnail>,
    pub download_url: Option<String>,
    pub user_permission: Option<Permission>,
    pub spaces: Option<Vec<String>>,
    pub shareable: Option<bool>,
    pub original_filename: Option<String>,
    pub quota_bytes_used: Option<String>,
    pub head_revision_id: Option<String>,
    pub self_link: Option<String>,
    pub file_size: Option<String>,
}

The metadata for a file.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

watch files (none)

empty trash files (none)

generate ids files (none)

copy files (request|response)

list files (none)

delete files (none)

patch files (request|response)

update files (request|response)

insert files (request|response)

untrash files (response)

trash files (response)

touch files (response)

get files (response)

export files (none)

Fields

mime_type

The MIME type of the file. This is only mutable on update when uploading new content. This field can be left blank, and the mimetype will be determined from the uploaded content's MIME type.

last_viewed_by_me_date

Last time this file was viewed by the user (formatted RFC 3339 timestamp).

app_data_contents

Whether this file is in the Application Data folder.

thumbnail_link

A short-lived link to the file's thumbnail. Typically lasts on the order of hours.

labels

A group of labels for the file.

indexable_text

Indexable text attributes for the file (can only be written)

explicitly_trashed

Whether this file has been explicitly trashed, as opposed to recursively trashed.

is_app_authorized

Whether the file was created or opened by the requesting app.

last_modifying_user_name

Name of the last user to modify this file.

writers_can_share

Whether writers can share the document with other users.

owned_by_me

Whether the file is owned by the current user.

shared_with_me_date

Time at which this file was shared with the user (formatted RFC 3339 timestamp).

sharing_user

User that shared the item with the current user, if available.

video_media_metadata

Metadata about video media. This will only be present for video types.

last_modifying_user

The last user to modify this file.

copyable

Whether the file can be copied by the current user.

folder_color_rgb

Folder color as an RGB hex string if the file is a folder. The list of supported colors is available in the folderColorPalette field of the About resource. If an unsupported color is specified, it will be changed to the closest color in the palette.

owner_names

Name(s) of the owner(s) of this file.

kind

The type of file. This is always drive#file.

id

The ID of the file.

web_view_link

A link only available on public folders for viewing their static web assets (HTML, CSS, JS, etc) via Google Drive's Website Hosting.

version

A monotonically increasing version number for the file. This reflects every change made to the file on the server, even those not visible to the requesting user.

parents

Collection of parent folders which contain this file. Setting this field will put the file in all of the provided folders. On insert, if no folders are provided, the file will be placed in the default root folder.

export_links

Links for exporting Google Docs to specific formats.

shared

Whether the file has been shared.

icon_link

A link to the file's icon.

full_file_extension

The full file extension; extracted from the title. May contain multiple concatenated extensions, such as "tar.gz". Removing an extension from the title does not clear this field; however, changing the extension on the title does update this field. This field is only populated for files with content stored in Drive; it is not populated for Google Docs or shortcut files.

open_with_links

A map of the id of each of the user's apps to a link to open this file with that app. Only populated when the drive.apps.readonly scope is used.

can_read_revisions

Whether the current user has read access to the Revisions resource of the file.

default_open_with_link

A link to open this file with the user's default app for this file. Only populated when the drive.apps.readonly scope is used.

description

A short description of the file.

web_content_link

A link for downloading the content of the file in a browser using cookie based authentication. In cases where the content is shared publicly, the content can be downloaded without any credentials.

editable

Whether the file can be edited by the current user.

marked_viewed_by_me_date

Deprecated.

embed_link

A link for embedding the file.

can_comment

Whether the current user can comment on the file.

file_extension

The final component of fullFileExtension with trailing text that does not appear to be part of the extension removed. This field is only populated for files with content stored in Drive; it is not populated for Google Docs or shortcut files.

modified_date

Last time this file was modified by anyone (formatted RFC 3339 timestamp). This is only mutable on update when the setModifiedDate parameter is set.

created_date

Create time for this file (formatted RFC 3339 timestamp).

properties

The list of properties.

md5_checksum

An MD5 checksum for the content of this file. This field is only populated for files with content stored in Drive; it is not populated for Google Docs or shortcut files.

permissions

The list of permissions for users with access to this file.

image_media_metadata

Metadata about image media. This will only be present for image types, and its contents will depend on what can be parsed from the image content.

owners

The owner(s) of this file.

etag

ETag of the file.

alternate_link

A link for opening the file in a relevant Google editor or viewer.

title

The title of this file.

modified_by_me_date

Last time this file was modified by the user (formatted RFC 3339 timestamp). Note that setting modifiedDate will also update the modifiedByMe date for the user which set the date.

thumbnail

Thumbnail for the file. Only accepted on upload and for files that are not already thumbnailed by Google.

download_url

no description provided

user_permission

The permissions for the authenticated user on this file.

spaces

The list of spaces which contain the file. Supported values are 'drive', 'appDataFolder' and 'photos'.

shareable

Whether the file's sharing settings can be modified by the current user.

original_filename

The original filename of the uploaded content if available, or else the original value of the title field. This is only available for files with binary content in Drive.

quota_bytes_used

The number of quota bytes used by this file.

head_revision_id

The ID of the file's head revision. This field is only populated for files with content stored in Drive; it is not populated for Google Docs or shortcut files.

self_link

A link back to this file.

file_size

The size of the file in bytes. This field is only populated for files with content stored in Drive; it is not populated for Google Docs or shortcut files.