Struct google_drive2::FileImageMediaMetadata

pub struct FileImageMediaMetadata {
    pub exposure_bias: Option<f32>,
    pub exposure_time: Option<f32>,
    pub max_aperture_value: Option<f32>,
    pub color_space: Option<String>,
    pub height: Option<i32>,
    pub lens: Option<String>,
    pub aperture: Option<f32>,
    pub date: Option<String>,
    pub rotation: Option<i32>,
    pub white_balance: Option<String>,
    pub camera_model: Option<String>,
    pub flash_used: Option<bool>,
    pub camera_make: Option<String>,
    pub focal_length: Option<f32>,
    pub exposure_mode: Option<String>,
    pub subject_distance: Option<i32>,
    pub width: Option<i32>,
    pub metering_mode: Option<String>,
    pub location: Option<FileImageMediaMetadataLocation>,
    pub iso_speed: Option<i32>,
    pub sensor: Option<String>,
}

Metadata about image media. This will only be present for image types, and its contents will depend on what can be parsed from the image content.

This type is not used in any activity, and only used as part of another schema.

Fields
exposure_bias: Option<f32>

The exposure bias of the photo (APEX value).

exposure_time: Option<f32>

The length of the exposure, in seconds.

max_aperture_value: Option<f32>

The smallest f-number of the lens at the focal length used to create the photo (APEX value).

color_space: Option<String>

The color space of the photo.

height: Option<i32>

The height of the image in pixels.

lens: Option<String>

The lens used to create the photo.

aperture: Option<f32>

The aperture used to create the photo (f-number).

date: Option<String>

The date and time the photo was taken (EXIF format timestamp).

rotation: Option<i32>

The rotation in clockwise degrees from the image's original orientation.

white_balance: Option<String>

The white balance mode used to create the photo.

camera_model: Option<String>

The model of the camera used to create the photo.

flash_used: Option<bool>

Whether a flash was used to create the photo.

camera_make: Option<String>

The make of the camera used to create the photo.

focal_length: Option<f32>

The focal length used to create the photo, in millimeters.

exposure_mode: Option<String>

The exposure mode used to create the photo.

subject_distance: Option<i32>

The distance to the subject of the photo, in meters.

width: Option<i32>

The width of the image in pixels.

metering_mode: Option<String>

The metering mode used to create the photo.

location: Option<FileImageMediaMetadataLocation>

Geographic location information stored in the image.

iso_speed: Option<i32>

The ISO speed used to create the photo.

sensor: Option<String>

The type of sensor used to create the photo.