Struct google_drive2::FileThumbnail

pub struct FileThumbnail {
    pub mime_type: Option<String>,
    pub image: Option<String>,
}

Thumbnail for the file. Only accepted on upload and for files that are not already thumbnailed by Google.

This type is not used in any activity, and only used as part of another schema.

Fields
mime_type: Option<String>

The MIME type of the thumbnail.

image: Option<String>

The URL-safe Base64 encoded bytes of the thumbnail image. It should conform to RFC 4648 section 5.