Struct google_drive2::Property

pub struct Property {
    pub self_link: Option<String>,
    pub kind: Option<String>,
    pub etag: Option<String>,
    pub visibility: Option<String>,
    pub key: Option<String>,
    pub value: Option<String>,
}

A key-value pair attached to a file that is either public or private to an application. The following limits apply to file properties:
- Maximum of 100 properties total per file - Maximum of 30 private properties per app - Maximum of 30 public properties - Maximum of 124 bytes size limit on (key + value) string in UTF-8 encoding for a single property.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert properties (request|response)

get properties (response)

patch properties (request|response)

update properties (request|response)

Fields

self_link

The link back to this property.

kind

This is always drive#property.

etag

ETag of the property.

visibility

The visibility of this property.

key

The key of this property.

value

The value of this property.